Remove firebase
