+++ examples/simple_send.rs
@@ -5,7 +5,6 @@ use web_push::*;
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
     let mut subscription_info_file = String::new();
-    let mut gcm_api_key: Option<String> = None;
     let mut vapid_private_key: Option<String> = None;
     let mut push_payload: Option<String> = None;
     let mut ttl: Option<u32> = None;
@@ -13,9 +12,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>
     {
         let mut ap = ArgumentParser::new();
         ap.set_description("A web push sender");
-
-        ap.refer(&mut gcm_api_key)
-            .add_option(&["-k", "--gcm_api_key"], StoreOption, "Google GCM API Key");
 
         ap.refer(&mut vapid_private_key).add_option(
             &["-v", "--vapid_key"],
@@ -48,10 +44,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>
 
     if let Some(ref payload) = push_payload {
         builder.set_payload(ContentEncoding::AesGcm, payload.as_bytes());
-    }
-
-    if let Some(ref gcm_key) = gcm_api_key {
-        builder.set_gcm_key(gcm_key);
     }
 
     if let Some(time) = ttl {
