+++ src/lib.rs
@@ -54,14 +54,13 @@ mod client;
 mod error;
 mod http_ece;
 mod message;
-mod response;
-mod services;
 mod vapid;
+mod web_push;
 
 #[cfg(feature = "client")]
 pub use crate::client::WebPushClient;
 pub use crate::error::WebPushError;
-pub use crate::response::read_response;
+pub use crate::web_push::read_response;
 
 pub use crate::message::{SubscriptionInfo, SubscriptionKeys, WebPushMessage, WebPushMessageBuilder, WebPushPayload};
