+++ src/message.rs
@@ -3,8 +3,8 @@ use std::fmt;
 use crate::{
     error::WebPushError,
     http_ece::{ContentEncoding, HttpEce},
-    services::*,
     vapid::VapidSignature,
+    web_push::build_request,
 };
 use base64;
 use http_types::{Request, Url};
@@ -45,12 +45,6 @@ impl SubscriptionInfo {
     }
 }
 
-#[derive(Debug, Clone, Copy)]
-pub enum WebPushService {
-    Firebase,
-    Autopush,
-}
-
 /// The push content payload, already in an encrypted form.
 #[derive(Debug, PartialEq)]
 pub struct WebPushPayload {
@@ -65,9 +59,6 @@ pub struct WebPushPayload {
 /// Everything needed to send a push notification to the user.
 #[derive(Debug)]
 pub struct WebPushMessage {
-    /// When not using VAPID, certain browsers need a Firebase account key for
-    /// sending a notification.
-    pub gcm_key: Option<String>,
     /// The endpoint URI where to send the payload.
     pub endpoint: Url,
     /// Time to live, how long the message should wait in the server if user is
@@ -75,26 +66,12 @@ pub struct WebPushMessage {
     pub ttl: u32,
     /// The encrypted request payload, if sending any data.
     pub payload: Option<WebPushPayload>,
-    /// The service type where to connect. Firebase when not using VAPID with
-    /// Chrome-based browsers. Data is in JSON format instead of binary.
-    pub service: WebPushService,
 }
 
 impl From<WebPushMessage> for Request {
     fn from(message: WebPushMessage) -> Self {
-        let service = message.service;
-
-        match service {
-            #[cfg(feature = "firebase")]
-            WebPushService::Firebase => {
-                trace!("Building firebase request");
-                firebase::build_request(message)
-            }
-            _ => {
-                trace!("Building autopush request");
-                autopush::build_request(message)
-            }
-        }
+        trace!("Building autopush request");
+        build_request(message)
     }
 }
 
@@ -113,7 +90,6 @@ impl<'a> fmt::Debug for WebPushPayloadBuilder<'a> {
 #[derive(Debug)]
 pub struct WebPushMessageBuilder<'a> {
     subscription_info: &'a SubscriptionInfo,
-    gcm_key: Option<&'a str>,
     payload: Option<WebPushPayloadBuilder<'a>>,
     ttl: u32,
     vapid_signature: Option<VapidSignature>,
@@ -128,7 +104,6 @@ impl<'a> WebPushMessageBuilder<'a> {
         Ok(WebPushMessageBuilder {
             subscription_info: subscription_info,
             ttl: 2_419_200,
-            gcm_key: None,
             payload: None,
             vapid_signature: None,
         })
@@ -139,11 +114,6 @@ impl<'a> WebPushMessageBuilder<'a> {
     /// delivery.
     pub fn set_ttl(&mut self, ttl: u32) {
         self.ttl = ttl;
-    }
-
-    /// For Google's push service, one must provide an API key from Firebase console.
-    pub fn set_gcm_key(&mut self, gcm_key: &'a str) {
-        self.gcm_key = Some(gcm_key);
     }
 
     /// Add a VAPID signature to the request. To be generated with the
@@ -166,15 +136,6 @@ impl<'a> WebPushMessageBuilder<'a> {
     pub fn build(self) -> Result<WebPushMessage, WebPushError> {
         let endpoint: Url = self.subscription_info.endpoint.parse().unwrap();
 
-        let service = match self.vapid_signature {
-            Some(_) => WebPushService::Autopush,
-            _ => match endpoint.host_str() {
-                Some("android.googleapis.com") => WebPushService::Firebase,
-                Some("fcm.googleapis.com") => WebPushService::Firebase,
-                _ => WebPushService::Autopush,
-            },
-        };
-
         if let Some(payload) = self.payload {
             let p256dh = base64::decode_config(&self.subscription_info.keys.p256dh, base64::URL_SAFE)?;
             let auth = base64::decode_config(&self.subscription_info.keys.auth, base64::URL_SAFE)?;
@@ -182,19 +143,15 @@ impl<'a> WebPushMessageBuilder<'a> {
             let http_ece = HttpEce::new(payload.encoding, &p256dh, &auth, self.vapid_signature);
 
             Ok(WebPushMessage {
-                gcm_key: self.gcm_key.map(|k| k.to_string()),
                 endpoint,
                 ttl: self.ttl,
                 payload: Some(http_ece.encrypt(payload.content)?),
-                service: service,
             })
         } else {
             Ok(WebPushMessage {
-                gcm_key: self.gcm_key.map(|k| k.to_string()),
                 endpoint: endpoint,
                 ttl: self.ttl,
                 payload: None,
-                service: service,
             })
         }
     }
