--- src/services/mod.rs
@@ -1,3 +1,0 @@
-pub mod autopush;
-#[cfg(feature = "firebase")]
-pub mod firebase;
